move privatekey functions into tls and fix up filtering for bulkcertificates
