+++ fastly/fixtures/platform_tls/create.yaml
@@ -2,8 +2,9 @@
 version: 1
 interactions:
 - request:
-    body: '{"Data":{"Type":"tls_private_key","Attributes":{"Key":"-----BEGIN PRIVATE
-      KEY-----\n...\n-----END PRIVATE KEY-----\n","Name":"My private key"}}}'
+    body: '{"data":{"type":"tls_bulk_certificate","attributes":{"cert_blob":"-----BEGIN
+      CERTIFICATE-----\n...\n-----END CERTIFICATE-----\n","intermediates_blob":"-----BEGIN
+      CERTIFICATE-----\n...\n-----END CERTIFICATE-----\n"},"relationships":{"tls_configurations":{"data":[{"type":"tls_configuration","id":"TLS_CONFIGURATION_ID"}]}}}}'
     form: {}
     headers:
       Accept:
@@ -12,10 +13,10 @@ interactions:
       - application/json
       User-Agent:
       - FastlyGo/1.7.2 (+github.com/fastly/go-fastly; go1.14)
-    url: https://api.fastly.com/tls/private_keys
+    url: https://api.fastly.com/tls/bulk/certificates
     method: POST
   response:
-    body: '{"data": {"id": "PRIVATE_KEY_ID","type": "tls_private_key","attributes": {"key_length": 2048,"key_type": "RSA","name": "My private key","created_at": "2019-02-01T12:12:12.000Z","replace": false,"public_key_sha1": "KEY_DIGEST"}}}'
+    body: '{"data": {"id": "CERTIFICATE_ID","type": "tls_bulk_certificate","attributes": {"not_after": "2019-06-06T18:14:32Z","not_before": "2018-06-06T18:14:32Z","created_at": "2018-06-06T18:14:32Z","updated_at": "2018-06-06T18:14:32Z","replace": false},"relationships": {"tls_configurations": {"data": [{"type": "tls_configuration","id": "TLS_CONFIGURATION_ID"}]},"tls_domains": {"data": [{"type": "tls_domain","id": "DOMAIN_NAME"}]}}}}'
     headers:
       Accept-Ranges:
       - bytes
