+++ fastly/fixtures/platform_tls/delete.yaml
@@ -1,16 +1,13 @@
 ---
 version: 1
-rwmutex: {}
 interactions:
 - request:
     body: ""
     form: {}
     headers:
-      Fastly-Key:
-      - xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
       User-Agent:
       - FastlyGo/1.7.2 (+github.com/fastly/go-fastly; go1.14)
-    url: https://api.fastly.com/tls/private_keys/PRIVATE_KEY_ID
+    url: https://api.fastly.com/tls/bulk/certificates/CERTIFICATE_ID
     method: DELETE
   response:
     body: ""
