+++ fastly/fixtures/tls/delete.yaml
@@ -1,13 +1,16 @@
 ---
 version: 1
+rwmutex: {}
 interactions:
 - request:
     body: ""
     form: {}
     headers:
+      Fastly-Key:
+      - xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
       User-Agent:
       - FastlyGo/1.7.2 (+github.com/fastly/go-fastly; go1.14)
-    url: https://api.fastly.com/tls/bulk/certificates/BULK_CERTIFICATE_ID
+    url: https://api.fastly.com/tls/private_keys/PRIVATE_KEY_ID
     method: DELETE
   response:
     body: ""
