+++ fastly/platform_tls_test.go
@@ -2,70 +2,6 @@ package fastly
 
 import "testing"
 
-func TestClient_PrivateKey(t *testing.T) {
-	t.Parallel()
-
-	fixtureBase := "platform_tls/"
-
-	// Create
-	var err error
-	var pk *PrivateKey
-	record(t, fixtureBase+"create", func(c *Client) {
-		pk, err = c.CreatePrivateKey(&CreatePrivateKeyInput{
-			Key:  "-----BEGIN PRIVATE KEY-----\n...\n-----END PRIVATE KEY-----\n",
-			Name: "My private key",
-		})
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// Ensure deleted
-	defer func() {
-		record(t, fixtureBase+"cleanup", func(c *Client) {
-			c.DeletePrivateKey(&DeletePrivateKeyInput{
-				ID: pk.ID,
-			})
-		})
-	}()
-
-	// List
-	var lpk []*PrivateKey
-	record(t, fixtureBase+"list", func(c *Client) {
-		lpk, err = c.ListPrivateKeys(&ListPrivateKeysInput{})
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-	if len(lpk) < 1 {
-		t.Errorf("bad privatekeys: %v", lpk)
-	}
-
-	// Get
-	var gpk *PrivateKey
-	record(t, fixtureBase+"get", func(c *Client) {
-		gpk, err = c.GetPrivateKey(&GetPrivateKeyInput{
-			ID: pk.ID,
-		})
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-	if pk.Name != gpk.Name {
-		t.Errorf("bad name: %q (%q)", pk.Name, gpk.Name)
-	}
-
-	// Delete
-	record(t, fixtureBase+"delete", func(c *Client) {
-		err = c.DeletePrivateKey(&DeletePrivateKeyInput{
-			ID: pk.ID,
-		})
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-}
-
 func TestClient_BulkCertificate(t *testing.T) {
 	t.Parallel()
 
@@ -74,7 +10,7 @@ func TestClient_BulkCertificate(t *testing.T) {
 	// Create
 	var err error
 	var bc *BulkCertificate
-	record(t, fixtureBase+"create_bulk", func(c *Client) {
+	record(t, fixtureBase+"create", func(c *Client) {
 		bc, err = c.CreateBulkCertificate(&CreateBulkCertificateInput{
 			CertBlob:          "-----BEGIN CERTIFICATE-----\n...\n-----END CERTIFICATE-----\n",
 			IntermediatesBlob: "-----BEGIN CERTIFICATE-----\n...\n-----END CERTIFICATE-----\n",
@@ -91,7 +27,7 @@ func TestClient_BulkCertificate(t *testing.T) {
 
 	// Ensure deleted
 	defer func() {
-		record(t, fixtureBase+"cleanup_bulk", func(c *Client) {
+		record(t, fixtureBase+"cleanup", func(c *Client) {
 			c.DeleteBulkCertificate(&DeleteBulkCertificateInput{
 				ID: bc.ID,
 			})
@@ -100,7 +36,7 @@ func TestClient_BulkCertificate(t *testing.T) {
 
 	// List
 	var lbc []*BulkCertificate
-	record(t, fixtureBase+"list_bulk", func(c *Client) {
+	record(t, fixtureBase+"list", func(c *Client) {
 		lbc, err = c.ListBulkCertificates(&ListBulkCertificatesInput{})
 	})
 	if err != nil {
@@ -112,7 +48,7 @@ func TestClient_BulkCertificate(t *testing.T) {
 
 	// Get
 	var gbc *BulkCertificate
-	record(t, fixtureBase+"get_bulk", func(c *Client) {
+	record(t, fixtureBase+"get", func(c *Client) {
 		gbc, err = c.GetBulkCertificate(&GetBulkCertificateInput{
 			ID: bc.ID,
 		})
@@ -126,7 +62,7 @@ func TestClient_BulkCertificate(t *testing.T) {
 
 	// Update
 	var ubc *BulkCertificate
-	record(t, fixtureBase+"update_bulk", func(c *Client) {
+	record(t, fixtureBase+"update", func(c *Client) {
 		ubc, err = c.UpdateBulkCertificate(&UpdateBulkCertificateInput{
 			ID:                "CERTIFICATE_ID",
 			CertBlob:          "-----BEGIN CERTIFICATE-----\n...\n-----END CERTIFICATE-----\n",
@@ -141,50 +77,9 @@ func TestClient_BulkCertificate(t *testing.T) {
 	}
 
 	// Delete
-	record(t, fixtureBase+"delete_bulk", func(c *Client) {
+	record(t, fixtureBase+"delete", func(c *Client) {
 		err = c.DeleteBulkCertificate(&DeleteBulkCertificateInput{
 			ID: bc.ID,
-		})
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-}
-
-func TestClient_ListPrivateKeys_validation(t *testing.T) {
-	t.Parallel()
-
-	var err error
-	record(t, "platform_tls/list", func(c *Client) {
-		_, err = c.ListPrivateKeys(&ListPrivateKeysInput{})
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-}
-
-func TestClient_GetPrivateKey_validation(t *testing.T) {
-	t.Parallel()
-
-	var err error
-	record(t, "platform_tls/get", func(c *Client) {
-		_, err = c.GetPrivateKey(&GetPrivateKeyInput{
-			ID: "PRIVATE_KEY_ID",
-		})
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-}
-
-func TestClient_CreatePrivateKey_validation(t *testing.T) {
-	t.Parallel()
-
-	var err error
-	record(t, "platform_tls/create", func(c *Client) {
-		_, err = c.CreatePrivateKey(&CreatePrivateKeyInput{
-			Key:  "-----BEGIN PRIVATE KEY-----\n...\n-----END PRIVATE KEY-----\n",
-			Name: "My private key",
 		})
 	})
 	if err != nil {
@@ -196,7 +91,7 @@ func TestClient_CreateBulkCertificate_validation(t *testing.T) {
 	t.Parallel()
 
 	var err error
-	record(t, "platform_tls/create_bulk", func(c *Client) {
+	record(t, "platform_tls/create", func(c *Client) {
 		_, err = c.CreateBulkCertificate(&CreateBulkCertificateInput{
 			CertBlob:          "-----BEGIN CERTIFICATE-----\n...\n-----END CERTIFICATE-----\n",
 			IntermediatesBlob: "-----BEGIN CERTIFICATE-----\n...\n-----END CERTIFICATE-----\n",
@@ -212,25 +107,11 @@ func TestClient_CreateBulkCertificate_validation(t *testing.T) {
 	}
 }
 
-func TestClient_DeletePrivateKey_validation(t *testing.T) {
-	t.Parallel()
-
-	var err error
-	record(t, "platform_tls/delete", func(c *Client) {
-		err = c.DeletePrivateKey(&DeletePrivateKeyInput{
-			ID: "PRIVATE_KEY_ID",
-		})
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-}
-
 func TestClient_DeleteBulkCertificate_validation(t *testing.T) {
 	t.Parallel()
 
 	var err error
-	record(t, "platform_tls/delete_bulk", func(c *Client) {
+	record(t, "platform_tls/delete", func(c *Client) {
 		err = c.DeleteBulkCertificate(&DeleteBulkCertificateInput{
 			ID: "CERTIFICATE_ID",
 		})
@@ -244,7 +125,7 @@ func TestClient_ListBulkCertificates_validation(t *testing.T) {
 	t.Parallel()
 
 	var err error
-	record(t, "platform_tls/list_bulk", func(c *Client) {
+	record(t, "platform_tls/list", func(c *Client) {
 		_, err = c.ListBulkCertificates(&ListBulkCertificatesInput{})
 	})
 	if err != nil {
@@ -256,7 +137,7 @@ func TestClient_GetBulkCertificate_validation(t *testing.T) {
 	t.Parallel()
 
 	var err error
-	record(t, "platform_tls/get_bulk", func(c *Client) {
+	record(t, "platform_tls/get", func(c *Client) {
 		_, err = c.GetBulkCertificate(&GetBulkCertificateInput{
 			ID: "CERTIFICATE_ID",
 		})
@@ -270,7 +151,7 @@ func TestClient_UpdateBulkCertificate_validation(t *testing.T) {
 	t.Parallel()
 
 	var err error
-	record(t, "platform_tls/update_bulk", func(c *Client) {
+	record(t, "platform_tls/update", func(c *Client) {
 		_, err = c.UpdateBulkCertificate(&UpdateBulkCertificateInput{
 			ID:                "CERTIFICATE_ID",
 			CertBlob:          "-----BEGIN CERTIFICATE-----\n...\n-----END CERTIFICATE-----\n",
